Fixed rubocop -a
