+++ app/controllers/password_resets_controller.rb
@@ -58,9 +58,9 @@ class PasswordResetsController < ApplicationController
   end
 
   def check_expiration
-    if @user.password_reset_expired?
-      flash[:danger] = 'Password reset has expired'
-      redirect_to new_password_reset_url
-    end
+    return unless @user.password_reset_expired?
+
+    flash[:danger] = 'Password reset has expired'
+    redirect_to new_password_reset_url
   end
 end
